--- docs/api_static/plasmapy.utils.roman.roman.rst
@@ -1,10 +1,0 @@
-:orphan:
-
-`plasmapy.utils.roman.roman`
-============================
-
-.. currentmodule:: plasmapy.utils.roman.roman
-
-.. automodapi::  plasmapy.utils.roman.roman
-   :include-all-objects:
-   :no-heading:
