+++ docs/api_static/plasmapy.utils.roman.rst
@@ -5,16 +5,6 @@
 
 .. currentmodule:: plasmapy.utils.roman
 
-.. automodule:: plasmapy.utils.roman
-   :noindex:
-
-Sub-Packages & Modules
-----------------------
-
-.. autosummary::
-
-   roman
-
 .. automodapi::  plasmapy.utils.roman
    :include-all-objects:
    :no-heading:
