+++ plasmapy/utils/__init__.py
@@ -10,8 +10,20 @@ __all__ = [
     "PlasmaPyWarning",
     "RelativityError",
     "RelativityWarning",
+    "call_string",
+    "decorators",
+    "exceptions",
+    "roman",
 ]
 
+from plasmapy.utils import (
+    datatype_factory_base,
+    decorators,
+    error_messages,
+    exceptions,
+    roman,
+)
+from plasmapy.utils.error_messages import call_string
 from plasmapy.utils.exceptions import (
     CouplingWarning,
     PhysicsError,
@@ -22,4 +34,9 @@ from plasmapy.utils.exceptions import (
     RelativityWarning,
 )
 
-from .error_messages import call_string
+try:
+    from plasmapy.utils import pytest_helpers
+except ModuleNotFoundError:
+    # pytest is not a hard dependency, so only import pytest_helpers is pytest
+    # is installed
+    pass
