--- plasmapy/utils/roman/__init__.py
@@ -1,10 +1,0 @@
-"""Utilities to convert integers to Roman numerals and vice versa."""
-
-from .roman import (
-    from_roman,
-    InvalidRomanNumeralError,
-    is_roman_numeral,
-    OutOfRangeError,
-    RomanError,
-    to_roman,
-)
